# 8298249: Excessive memory allocation in CipherInputStream AEAD decryption