Improves `CipherInputStream` with `AES-GCM` significantly. This problem is seen in some of our services.

There is a related [JDK-8298865](https://bugs.openjdk.org/browse/JDK-8298865), which I would backport separately.

Additional testing:
 - [x] MacOS AArch64 server fastdebug, `jdk_security`
 - [ ] Linux x86_64 server fastdebug, `all`
 - [x] Ad-hoc benchmarks 

```
Benchmark                        (dataSize)  (keyLength)    Mode  Cnt       Score     Error  Units

# Before
AESGCMCipherInputStream.decrypt       16384          128   thrpt   40   35547.250 ± 154.108  ops/s
AESGCMCipherInputStream.decrypt     1048576          128   thrpt   40      17.297 ±   0.043  ops/s
AESGCMCipherInputStream.encrypt       16384          128   thrpt   40  150791.244 ± 656.878  ops/s
AESGCMCipherInputStream.encrypt     1048576          128   thrpt   40    2561.195 ±  27.701  ops/s

# After
AESGCMCipherInputStream.decrypt       16384          128   thrpt   40   65815.523 ± 156.983  ops/s
AESGCMCipherInputStream.decrypt     1048576          128   thrpt   40     786.680 ± 298.170  ops/s
AESGCMCipherInputStream.encrypt       16384          128   thrpt   40  149412.181 ± 431.304  ops/s
AESGCMCipherInputStream.encrypt     1048576          128   thrpt   40    2581.110 ±  23.931  ops/s
```

<!-- Anything below this marker will be automatically updated, please do not edit manually! -->
---------
### Progress
- [x] Change must not contain extraneous whitespace
- [x] Commit message must refer to an issue
- [ ] [JDK-8298249](https://bugs.openjdk.org/browse/JDK-8298249) needs maintainer approval

### Issue
 * [JDK-8298249](https://bugs.openjdk.org/browse/JDK-8298249): Excessive memory allocation in CipherInputStream AEAD decryption (**Bug** - P4)


### Reviewing
<details><summary>Using <code>git</code></summary>

Checkout this PR locally: \
`$ git fetch https://git.openjdk.org/jdk17u-dev.git pull/2393/head:pull/2393` \
`$ git checkout pull/2393`

Update a local copy of the PR: \
`$ git checkout pull/2393` \
`$ git pull https://git.openjdk.org/jdk17u-dev.git pull/2393/head`

</details>
<details><summary>Using Skara CLI tools</summary>

Checkout this PR locally: \
`$ git pr checkout 2393`

View PR using the GUI difftool: \
`$ git pr show -t 2393`

</details>
<details><summary>Using diff file</summary>

Download this PR as a diff file: \
<a href="https://git.openjdk.org/jdk17u-dev/pull/2393.diff">https://git.openjdk.org/jdk17u-dev/pull/2393.diff</a>

</details>


### Webrev
[Link to Webrev Comment](https://git.openjdk.org/jdk17u-dev/pull/2393#issuecomment-2049858320)