@shipilev 2024-04-11
Cannot do it easily for the reasons stated in the bug: https://bugs.openjdk.org/browse/JDK-8298249?focusedId=14664713#comment-14664713